+++ src/midi/index.ts
@@ -1,6 +1,7 @@
 export * from "./cc-rate-limit";
 export * from "./connection";
 export * from "./errors";
+export * from "./pipelined-read";
 export * from "./ports";
 export * from "./reassembly";
 export * from "./request-response";